html/browsers/browsing-the-web/navigating-across-documents/018-form-action-javascript-url-returns-script-element.html

<!doctype html>
<link rel="help" href="https://html.spec.whatwg.org/multipage/browsing-the-web.html#navigating-across-documents">
<link rel="help" href="https://html.spec.whatwg.org/multipage/links.html#following-hyperlinks">
<link rel="help" href="https://html.spec.whatwg.org/multipage/browsing-the-web.html#javascript-protocol">
<link rel="help" href="https://html.spec.whatwg.org/multipage/form-control-infrastructure.html#concept-form-submit">
<link rel="help" href="https://html.spec.whatwg.org/multipage/form-control-infrastructure.html#submit-get-action">
<link rel="help" href="https://html.spec.whatwg.org/multipage/form-control-infrastructure.html#plan-to-navigate">
<link rel="help" href="https://html.spec.whatwg.org/multipage/form-control-infrastructure.html#planned-navigation">
<link rel="help" href="https://html.spec.whatwg.org/multipage/webappapis.html#queue-a-task">
<link rel="help" href="https://html.spec.whatwg.org/multipage/webappapis.html#dom-manipulation-task-source">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<iframe src=""></iframe>
<script>
const runner = location.href.slice(location.href.lastIndexOf("/") + 1, location.href.length)
                  .replace(".html", "-runner.html");

async_test(function(test) {

  window.verifyNavigation = test.step_func_done(function(signal) {
    assert_equals(signal, true, "The navigation via 'javascript: url' code did not complete as expected.");
  });

  document.querySelector("iframe").src = runner;

}, `A navigation via 'javascript: url' code (in: ${runner}) must be evaluated async; to pass this test, that code must be evaluated and the resulting string written to the iframe body as the contents of a script element. There is no default href navigation in the link.`);
</script>
